backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
psycopg[binary]
passlib[bcrypt]
python-jose
PyJWT
python-dotenv
python-multipart
redis
requests
jinja2
phonenumbers
aiosmtplib
email-validator
python-socketio
qrcode
pillow
twilio